extracted query methods
